Move redux store generation logic to `index.ts`
